TASK-006: Move misplaced utilities to architecturally correct locations
Move utility functions from cli/utils.py (780 lines) to purpose-specific
modules aligned with the dependency hierarchy, fixing the web→cli Level 2→3
architecture violation.

New files:
- utils/process.py — is_process_running, kill_process (Level 0)
- config/helpers.py — find_project_root, get_config, load_config, save_config,
  check_unreferenced_custom_sources, format_unreferenced_sources_warning (Level 0)
- platform/watcher_lifecycle.py — watcher start/stop/status (Level 2)
- cli/helpers/port_manager.py — check_port_in_use, get_process_using_port (Level 3)
- cli/helpers/process_manager.py — FastAPI server PID management (Level 3)

Key changes:
- cli/utils.py shrinks from 780→129 lines (display helpers only)
- web/routes/health.py now imports from platform/ instead of cli/ (fixes violation)
- stop_file_watcher loses verbose param (callers print their own messages)
- config/loader.py trimmed — convenience functions moved to config/helpers.py
- All 201 tests pass, ruff clean
