+++ dango/cli/helpers/__init__.py
@@ -1,0 +1,6 @@
+"""dango/cli/helpers/__init__.py
+
+CLI helper subpackage. Import directly from submodules:
+    from dango.cli.helpers.port_manager import check_port_in_use
+    from dango.cli.helpers.process_manager import start_fastapi_server
+"""
